refactor(dashboard): extract war-room-metrics from war-room.ts
Split war-room.ts (942 LOC) into war-room.ts + war-room-metrics.ts.
Secondary rendering components extracted to reduce main file size.
